feat(Pit): 가장 멀리간 차를 알 수 있다
