+++ src/main/java/racingcar/domain/Car.java
@@ -17,8 +17,12 @@ public class Car {
         }
     }
 
-    public boolean hasHigherPositionThan(Car another) {
-        return this.position >= another.position;
+    public boolean isFartherThan(Car other) {
+        return position > other.position;
+    }
+
+    public boolean hasSamePositionWith(Car other) {
+        return position == other.position;
     }
 
     public String getName() {
